+++ tomcat/src/main/java/nextstep/jwp/controller/ViewController.java
@@ -3,10 +3,9 @@ package nextstep.jwp.controller;
 import java.io.IOException;
 import java.util.Objects;
 import javassist.NotFoundException;
+import org.apache.catalina.servlet.handler.ServletResponse;
 import org.apache.catalina.servlet.util.ResourceReader;
 import org.apache.coyote.http11.request.Request;
-import org.apache.coyote.http11.response.Response;
-import org.apache.coyote.http11.response.Response.ServletResponse;
 import org.apache.coyote.http11.session.SessionManager.Session;
 
 public class ViewController implements Controller {
@@ -25,10 +24,10 @@ public class ViewController implements Controller {
     private static ServletResponse responseByLogin(final Session session, final String redirectUri)
             throws NotFoundException, IOException {
         if (isLoginUser(session)) {
-            return Response.redirect("index.html");
+            return ServletResponse.redirect("index.html");
         }
 
-        return Response.staticResource(ResourceReader.read(redirectUri));
+        return ServletResponse.staticResource(ResourceReader.read(redirectUri));
     }
 
     private static boolean isLoginUser(final Session session) {
