+++ tomcat/src/main/java/org/apache/catalina/servlet/handler/RequestHandler.java
@@ -1,7 +1,6 @@
 package org.apache.catalina.servlet.handler;
 
 import org.apache.coyote.http11.request.Request;
-import org.apache.coyote.http11.response.Response.ServletResponse;
 
 public interface RequestHandler {
