+++ tomcat/src/main/java/org/apache/catalina/servlet/handler/Servlet.java
@@ -1,12 +1,11 @@
 package org.apache.catalina.servlet.handler;
 
+import static org.apache.catalina.servlet.handler.ServletResponse.methodNotAllowed;
 import static org.apache.coyote.http11.common.Method.GET;
 import static org.apache.coyote.http11.common.Method.POST;
 import static org.apache.coyote.http11.common.Status.METHOD_NOT_ALLOWED;
 
 import org.apache.coyote.http11.request.Request;
-import org.apache.coyote.http11.response.Response;
-import org.apache.coyote.http11.response.Response.ServletResponse;
 
 public abstract class Servlet implements RequestHandler {
 
@@ -20,12 +19,13 @@ public abstract class Servlet implements RequestHandler {
     public void service(final Request request, final ServletResponse response) throws Exception {
         if (request.getMethod() == GET) {
             doGet(request, response);
+            return;
         }
         if (request.getMethod() == POST) {
             doPost(request, response);
+            return;
         }
-        Response.methodNotAllowed()
-                .build();
+        methodNotAllowed().build();
     }
 
     public String getMappingPath() {
